+++ example/android/app/src/main/kotlin/com/example/curved_app_bar_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.curved_app_bar_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()