+++ lib/src/curved_app_bar.dart
@@ -1,3 +1,4 @@
 export 'clippers/curved_app_bar_clipper.dart';
 export 'enums/curved_app_bar_shape.dart';
+export 'models/curved_app_bar_action.dart';
 export 'widgets/curved_app_bar.dart';